Async modbus client

+++ src/modbus/modbus_server.h
@@ -3,9 +3,6 @@
 
 #pragma once
 
-#include <stdbool.h>
-#include <stdint.h>
-
 #include "modbus_common.h"
 
 #define MB_SERVER_RECEIVE_TIMEOUT 100
@@ -20,9 +17,9 @@ struct mb_server_cb {
   enum mb_result (*write_multiple_coils)(uint16_t start, uint8_t* data, uint16_t count);
   enum mb_result (*write_multiple_registers)(uint16_t start, uint16_t* data, uint16_t count);
 
-  void (*pass_thru)(uint8_t* data, uint32_t len);
+  void (*pass_thru)(uint8_t* data, size_t len);
 
-  void (*tx)(uint8_t* data, uint32_t len);
+  void (*tx)(uint8_t* data, size_t len);
   uint32_t (*get_tick_ms)(void);
 };
